change int* triangleSlidesPtr to float* triangleSlidesPtr to have variables properly enter verifyTriangle function

+++ PolygonChecker/main.c
@@ -18,7 +18,7 @@ int main() {
 		case 1:
 			printf_s("Triangle selected.\n");
 			float triangleSides[3] = { 0.f, 0.f, 0.f };
-			int* triangleSidesPtr = getTriangleSides(triangleSides);
+			float* triangleSidesPtr = getTriangleSides(triangleSides);
 			bool verify = verifyTriangle(triangleSidesPtr[0], triangleSidesPtr[1], triangleSidesPtr[2]);
 			//printf_s("! %d\n", triangleSidesPtr[0]);
 
@@ -26,10 +26,9 @@ int main() {
 
 				char* result = analyzeTriangle(triangleSidesPtr[0], triangleSidesPtr[1], triangleSidesPtr[2]);
 				printf_s("%s\n", result);
-
 			}
-
 			break;
+
 		case 2:
 			printf("Hello World");
 		case 0:
@@ -94,19 +93,16 @@ int* getTriangleSides(int* triangleSides) {
 					valid = true;
 				}
 				else{
-
 					printf("Triangle is not valid.\n");
 					valid = false;
 				}
 			}
 			else{
-
 				printf("Triangle is not valid.\n");
 				valid = false;
 			}
 		}
 		else{
-
 			printf("Triangle is not valid.\n");
 			valid = false;
 		}
